\name{ccgplot}
\alias{ccgplot}
\title{ccgplot}
\usage{
  ccgplot(ccg, ...)
}
\arguments{
  \item{ccg}{A composite graph object}

  \item{...}{Parameters to be passed to plot.igraph for g1
  and g2}
}
\description{
  Plot a collaborative composite graph
}
\details{
  Plot a collaborative composite graph
}
\examples{
plot(ercg(20, 0.5))
}

