\name{ccgraph}
\alias{ccgraph}
\alias{ccgraph.default}
\title{ccgraph}
\arguments{
  \item{cg}{A compgraph object}

  \item{...}{Currently ignored}
}
\value{
  A ccgraph object, or an error
}
\description{
  Instantiate a ccgraph object
}
\details{
  Class function for collaborative composite graphs
}
\examples{
showClass("GameDayPlays")
}
\author{
  Ben Baumer
}

