\name{plot.compgraph}
\alias{plot.compgraph}
\title{plot.compgraph}
\usage{
  plot.compgraph(G, hilite.cbtime = TRUE, ...)
}
\arguments{
  \item{G}{A composite graph object}

  \item{hilite.cbtime}{A boolean value}

  \item{...}{Parameters to be passed to plot.igraph for g1
  and g2}
}
\description{
  Plot a composite graph
}
\details{
  Plot a composite graph
}
\examples{
plot(ercg(20, 0.5))
}

